LU-18740 mgs: size_t in contain_valid_fsname()
to fix a build warning with gcc 11.5.0 (Rocky 9.3):

lustre/mgs/mgs_llog.c:4995:13: error: '__builtin_memcmp_eq'
 specified bound [18446744071562067968, 0] exceeds maximum
 object size 9223372036854775807 [-Werror=stringop-overread]
 4995 |         if (memcmp(buf, fsname, namelen) != 0)
      |             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58142
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/mgs/mgs_llog.c
@@ -4987,7 +4987,7 @@ struct mgs_lcfg_fork_data {
 };
 
 static bool contain_valid_fsname(char *buf, const char *fsname,
-				 int buflen, int namelen)
+				 size_t buflen, size_t namelen)
 {
 	if (buflen < namelen)
 		return false;
